Refactor experiments
- adds bar formatter to experiment evals
- adds optional name parameter to many evaluators
